Added print_*_deps targets to allow viewing and debugging dependencies for various targets.
At present, it looks like the 'goog.Uri' dependency is breaking dependency
calculation for chess_fen_in.js (although compilation works and the result
seems to be correct), so the way to compile it when its deps change is to force
it via:

% touch src/chess_fen_in.js
% make

because the "src/chess_fen_in.js" dependency works, but the auto-calculated
dependencies are empty. This does not affect other games, whose demos are not
currently requiring the 'goog.Uri' package.